+++ src/main/java/com/rastatech/secretrasta/controller/WishesController.java
@@ -1,14 +1,18 @@
 package com.rastatech.secretrasta.controller;
 
-import com.rastatech.secretrasta.dto.*;
-import com.rastatech.secretrasta.model.CommentEntity;
+import com.rastatech.secretrasta.dto.UpdateWishRequest;
+import com.rastatech.secretrasta.dto.WishRequest;
+import com.rastatech.secretrasta.dto.WishResponse;
 import com.rastatech.secretrasta.model.WishEntity;
+import com.rastatech.secretrasta.repository.WishRepository;
 import com.rastatech.secretrasta.service.UserService;
 import com.rastatech.secretrasta.service.WishService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -52,12 +56,18 @@ public class WishesController {
 
     @PutMapping("/{wish_id}")
     public void updateWish(@PathVariable("wish_id") Long wishId,
-                              @Valid @RequestBody UpdateWishRequest wish) {
+                           @Valid @RequestBody UpdateWishRequest wish,
+                           Authentication auth) {
+        if (!auth.getPrincipal().equals(wishService.fetchWish(wishId).getUser().getUsername()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         wishService.updateWish(wishId, wish);
     }
 
     @DeleteMapping("/{wish_id}")
-    public void deleteWish(@PathVariable("wish_id") Long wishId) {
+    public void deleteWish(@PathVariable("wish_id") Long wishId,
+                           Authentication auth) {
+        if (!auth.getPrincipal().equals(wishService.fetchWish(wishId).getUser().getUsername()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         wishService.deleteWish(wishId);
     }
